pyproject.toml
[project]
name = "leverx-homework-3"
version = "0.1.0"
description = "MySQL loader and reports"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "LeverX Homework" }
]
dependencies = [
  "mysql-connector-python==9.0.0",
  "python-dotenv==1.0.1",
  "ijson==3.3.0",
]

[tool.uv]
# You can add per-platform constraints or indexes here if needed.
